Changed filters in generator.py from Tensor to int

+++ src/model/discriminator.py
@@ -20,7 +20,7 @@ def ACGANDiscriminator(input_tensor,
         Tensor[None, num_classes]: A tensor representing the classification for the image.
     """
     with tf.variable_scope('Discriminator', reuse=tf.AUTO_REUSE):
-        num_convolutions = math.log(output_dim, 2) - 1
+        num_convolutions = int(math.log(output_dim, 2) - 1)
 
         with tf.variable_scope('layers'):
             for i in range(num_convolutions):
